==> 01_data_prep.ipynb <==
import pandas as pd
import numpy as np

# Prep Golden Mapping

def remove_empty(x):
    result = sorted(list(set(x)))
    try:
        result.remove('')
    except:
        pass
    return result

golden = pd.read_csv('./2020_repologue_data/golden_set_v3.csv')

golden['alias'] = golden['alias'].replace(np.nan, '', regex=True)
golden['alias'] = golden['alias'].apply(lambda x: x+',')
golden['our_sub_topics'] = golden['our_sub_topics'].replace(np.nan, '', regex=True)

golden['free'] = golden['alias']+golden['our_sub_topics']
golden = golden[golden.free != ","]
golden['free'] = golden['free'].replace(np.nan, '', regex=True)
golden['free'] = golden['free'].apply(lambda x: x.split(','))
golden['free'] = golden['free'].apply(lambda x: sorted(list(set(x))))
golden['free'] = golden['free'].apply(lambda x: remove_empty(x))
golden['free_str'] = golden['free'].apply(lambda x: ','.join(x))

golden.rename(columns={'Topics':'topic'}, inplace=True)
golden.drop(['alias','our_sub_topics'], axis=1, inplace=True)

frees = sorted(list(set([x for row in list(golden.free) for x in row])))

# Prep Mapped Data

df = pd.read_csv('./2020_repologue_data/repos_152k_top228_topics_compressed_2.csv')

df['featured_topics'] = df['featured_topics'].replace(np.nan, '', regex=True)

df['original_topics'] = df['original_topics'].apply(lambda x: x.split(','))
df['featured_topics'] = df['featured_topics'].apply(lambda x: x.split(','))
df['augmented_top228topics'] = df['augmented_top228topics'].apply(lambda x: x.split(','))

df['featured_topics'] = df['featured_topics'].apply(lambda x: [] if x==[""] else x)

original_topics = sorted(list(set([x for row in list(df.original_topics) for x in row])))
featured_topics = sorted(list(set([x for row in list(df.featured_topics) for x in row])))
augmented_top228topics = sorted(list(set([x for row in list(df.augmented_top228topics) for x in row])))

# Map originals based on the Golden Mapping

def get_full_mapping(x,frees=frees,golden=golden):
    result = []
    for item in x:
        if item in frees:
            result.append(golden[golden['free_str'].str.contains(item)]['topic'].values[0])
    return result

df.insert(4,'mapped_topics', df['original_topics'].apply(get_full_mapping))

df['mapped_topics'] = df['featured_topics']+df['augmented_top228topics']+df['mapped_topics']
df['mapped_topics'] = df['mapped_topics'].apply(lambda x: sorted(list(set(x))))

mapped_topics = sorted(list(set([x for row in list(df.mapped_topics) for x in row])))

# Reports

df.insert(6, 'len_original_topics', df['original_topics'].apply(lambda x : len(x)))
df.insert(7, 'len_featured_topics', df['featured_topics'].apply(lambda x : len(x)))
df.insert(8, 'len_augmented_top228topics', df['augmented_top228topics'].apply(lambda x : len(x)))
df.insert(9, 'len_mapped_topics', df['mapped_topics'].apply(lambda x : len(x)))

print('Topic Count')
print(f"original_topics\t\t{len(original_topics)}\nfeatured_topics\t\t{len(featured_topics)}\
        \naugmented_top228topics\t{len(augmented_top228topics)}\nmapped_topics\t\t{len(mapped_topics)}")

print()

print('Averages Number of Topics Per Repository')
print(f"original_topics\t\t{df['len_original_topics'].mean()}\nfeatured_topics\t\t{df['len_featured_topics'].mean()}\
        \naugmented_top228topics\t{df['len_augmented_top228topics'].mean()}\nmapped_topics\t\t{df['len_mapped_topics'].mean()}")

# Saving Prepped Data

to_drop = ['readme_raw']

df.drop(to_drop, axis=1, inplace=True)

df.original_topics = df.original_topics.apply(lambda x: ','.join(x))
df.featured_topics = df.featured_topics.apply(lambda x: ','.join(x))
df.augmented_top228topics = df.augmented_top228topics.apply(lambda x: ','.join(x))
df.mapped_topics = df.mapped_topics.apply(lambda x: ','.join(x))

df.to_csv('./2020_repologue_data/prepped_data.csv', index=None)

df = pd.read_csv('./2020_repologue_data/prepped_data.csv')

df
