==> 04_ml_results_comparison.ipynb <==
import pickle
import pandas as pd
import numpy as np
from sklearn.feature_extraction.text import TfidfTransformer, TfidfVectorizer
from sklearn.multiclass import OneVsRestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.naive_bayes import MultinomialNB
from sklearn.metrics import classification_report


path = './learning_data/'

# Load Test Data

with open(path+'X_test_raw.pkl','rb') as f:
    X_test_raw = pickle.load(f)
with open(path+'y_test.pkl','rb') as f:
    y_test = pickle.load(f)
with open(path+'topics.pkl','rb') as f:
    topics = pickle.load(f)

# Load Models

# LR

with open(path+'tfidf_vectorizer.pkl','rb') as f:
    tfidf_vectorizer_lr = pickle.load(f)
with open('./learning_data/ovr_clf.pkl','rb') as f:
    ovr_clf = pickle.load(f)

# MNB

with open(path+'tfidf_vectorizer_mnb.pkl','rb') as f:
    tfidf_vectorizer_mnb = pickle.load(f)
with open(path+'tfidf_transformer_mnb.pkl','rb') as f:
    tfidf_transformer_mnb = pickle.load(f)
with open(path+'mnb_clf.pkl','rb') as f:
    mnb_clf = pickle.load(f)

# Helper function to get the top-k results

def get_top_k(probs, k):
    results = np.zeros_like(probs)
    for i in range(probs.shape[0]):
        indexes = probs[i,:].argsort()[-k:][::-1]
        results[i,indexes] = 1.0
    return results

# Results

k = 10

# LR

X_test_lr = tfidf_vectorizer_lr.transform(X_test_raw.input.values)
lr_predictions = ovr_clf.predict_proba(X_test_lr)
lr_results = get_top_k(lr_predictions, k)

print('\n'+'*'*25+' LR Results:')
print(classification_report(y_test, lr_results, target_names=topics, digits=4, zero_division=0))

# MNB

X_test_mnb = tfidf_transformer_mnb.transform(tfidf_vectorizer_mnb.transform(X_test_raw.readme_processed.values))
mnb_predictions = mnb_clf.predict_proba(X_test_mnb)
mnb_results = get_top_k(mnb_predictions, k)

print('\n'+'*'*25+' MNB Results:')
print(classification_report(y_test, mnb_results, target_names=topics, digits=4, zero_division=0))
