==> 05_extract_sample_index_for_eval.ipynb <==
import pickle
import numpy as np
from random import sample
import os

# Extract k sample indexes and save to ./results/eval_sample_indexes.pkl as a list

k = 75

with open('./learning_data/y_test.pkl','rb') as f:
    y_test = pickle.load(f)

all_sample_count = y_test.shape[0]

indexes = sample(range(all_sample_count), k)

try:
    os.makedirs('./results/')
except:
    pass

with open('./results/eval_sample_indexes.pkl','wb') as f:
    pickle.dump(indexes, f)
